Add i18n test of components
