+++ package.json
@@ -56,6 +56,7 @@
     "@testing-library/react": "^14.0.0",
     "@types/jest": "^27.5.2",
     "@types/lodash-es": "^4.17.8",
+    "@types/react-test-renderer": "^18.0.1",
     "@types/styled-components": "^5.1.26",
     "@typescript-eslint/eslint-plugin": "^5.61.0",
     "@typescript-eslint/parser": "^5.61.0",
